pyproject.toml
[tool.poetry]
name = "langchain_project"
version = "0.1.0"
description = ""
authors = ["Jerry <jerry8g@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.3.28"
langchain-community = "^0.3.27"
dotenv = "^0.9.9"
tiktoken = "^0.9.0"
redis = "^6.2.0"
sqlalchemy-utils = "^0.41.2"
gptcache = "^0.1.44"
openai = "^1.97.0"
faiss-cpu = "^1.11.0.post1"
langchain-huggingface = "^0.3.0"
transformers = "^4.53.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
